paddle-ocr/requirements.txt
paddlepaddle
paddleocr
opencv-python==4.2.0.32
opencv-python-headless==4.2.0.32
fastapi[all]
uvicorn
Pillow
